ExploringLLMs/OpenAI/liveInference_with_openAI/requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
openai>=1.0.0
python-dotenv>=1.0.0
pillow>=10.0.0 